working on domain

+++ e-bookshop/Catalog/Catalog.Application/DTOs/Books/UpdatePriceBookDto.cs
@@ -3,7 +3,7 @@
     public class UpdatePriceBookDto
     {
         public Guid Id { get; set;}
-        public double NewPrice { get; set;}
+        public decimal NewPrice { get; set;}
         public required string NewCurrency { get; set; }
 
     }
